optimize module dependency graph, so that the UnidocLinker no longer depends on MarkdownAST
