--- Sources/MarkdownLinking/exports.swift
@@ -1,1 +1,0 @@
-@_exported import enum MarkdownABI.Markdown
