+++ Sources/MarkdownSemantics/Sema/Markdown.SemanticAnalyzer.swift
@@ -1,5 +1,4 @@
 import MarkdownAST
-import MarkdownLinking
 import SourceDiagnostics
 
 extension Markdown
